Handle non-final errors when sending transactions
Median time-past is used for lock-time calculations which means
that the "network time" might not be the exact same as the actual
time. This can result in a non-final error when trying to spend
a transaction that should be unlocked according to the actual time.
This commit handles these errors and explains them to the user.

More info can be found here:
<https://github.com/bitcoin/bips/blob/master/bip-0113.mediawiki>