Make sure the DatePicker returns the correct date
This commit work arounds a bug that causes the DatePicker to return
an incorrect date. This is a known bug with the DatePickerIOS component
(and perhaps even in iOS).

The workaround is to use a modified component that exposes a getter
to get the date instead of relying on events/callbacks.

Read more about the issue here:
https://github.com/facebook/react-native/issues/4787
https://github.com/facebook/react-native/issues/8169