Show error message if selected date has past or is too far
Previously it was possible to select a past date. This caused
problems when the date was before the date the BitHodl Service
start scanning at. This fix shows an error message if the selected
date is in the past or if it's too far in the future.

It doesn't use the `maximumDate` and `minimumDate` properties of
the DatePicker anymore since they automatically adjusted the date
which can be annoying for the user.