Automatically focus the next recovery phrase input
When the user has entered the correct recovery phrase word,
automatically focus the next for a better user expecience.

Only do this when the words are known (confirming the
recovery phrase). When recovering a wallet, both 'man'
and 'mansion' are correct words, so tabbing on 'man'
would make it impossible to enter 'mansion'.